Tasks Update Backend and Dashboard API update

+++ src/main/java/team/placeholder/internalprojectsmanagementsystem/dto/uidto/PlanManMonthDto.java
@@ -6,6 +6,6 @@ import java.util.Map;
 
 @Data
 public class PlanManMonthDto {
-    private long planManMonthHours;
+    private Double planManMonthHours;
     private String monthName;
 }
